Move getSubset to common pdf/ps backend
