+++ lib/matplotlib/backends/_backend_pdf_ps.py
@@ -2,7 +2,10 @@
 Common functionality between the PDF and PS backends.
 """
 
+from io import BytesIO
 import functools
+
+from fontTools import subset
 
 import matplotlib as mpl
 from .. import font_manager, ft2font
@@ -14,6 +17,28 @@ from ..backend_bases import RendererBase
 def _cached_get_afm_from_fname(fname):
     with open(fname, "rb") as fh:
         return AFM(fh)
+
+
+def getSubset(self, fontfile, characters):
+    """
+    Subset a TTF font
+
+    Reads the named fontfile and restricts the font to the characters.
+    Returns a serialization of the subset font as bytes.
+    """
+
+    options = subset.Options(glyph_names=True, recommended_glyphs=True)
+    options.drop_tables += ['FFTM']
+    font = subset.load_font(fontfile, options)
+    try:
+        subsetter = subset.Subsetter(options=options)
+        subsetter.populate(text=characters)
+        subsetter.subset(font)
+        fh = BytesIO()
+        font.save(fh, reorderTables=False)
+        return fh.getvalue()
+    finally:
+        font.close()
 
 
 class CharacterTracker:
